refactor business logic from repos (scenario, pipeline, task) to entities
