+++ tests/core/task/test_task_repository.py
@@ -51,20 +51,6 @@ task = Task(
     version="latest",
 )
 
-task_model = _TaskModel(
-    id="id",
-    owner_id="owner_id",
-    parent_ids=["parent_id"],
-    config_id="config_id",
-    input_ids=["dn_id"],
-    function_name=print.__name__,
-    function_module=print.__module__,
-    output_ids=[],
-    version="latest",
-    skippable=False,
-    properties={},
-)
-
 
 class TestTaskRepository:
     def test_save_and_load(self, tmpdir):
@@ -76,16 +62,6 @@ class TestTaskRepository:
         _DataManager._set(data_node)
         t = repository.load("id")
         assert t.id == task.id
-        assert len(t.input) == 1
-
-    def test_from_and_to_model(self):
-        repository = _TaskRepositoryFactory._build_repository()  # type: ignore
-        assert repository._to_model(task) == task_model
-        with pytest.raises(NonExistingDataNode):
-            repository._from_model(task_model)
-        _DataManager._set(data_node)
-        t = repository._from_model(task_model)
-        assert isinstance(t, Task)
         assert len(t.input) == 1
 
     def test_save_and_load_with_sql_repo(self, tmpdir):
@@ -100,18 +76,4 @@ class TestTaskRepository:
         _DataManager._set(data_node)
         t = repository.load("id")
         assert t.id == task.id
-        assert len(t.input) == 1
-
-    def test_from_and_to_model_with_sql_repo(self):
-        Config.configure_global_app(repository_type="sql")
-
-        _DataManagerFactory._build_manager()._delete_all()
-        repository = _TaskRepositoryFactory._build_repository()  # type: ignore
-
-        assert repository._to_model(task) == task_model
-        with pytest.raises(NonExistingDataNode):
-            repository._from_model(task_model)
-        _DataManager._set(data_node)
-        t = repository._from_model(task_model)
-        assert isinstance(t, Task)
         assert len(t.input) == 1
